% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorapp.R
\name{colordot}
\alias{colordot}
\title{function to plot a dot with a given color balance}
\usage{
colordot(r, g, b)
}
\arguments{
\item{r}{numeric(1) should lie in [0,1], red proportion}

\item{g}{numeric(1) should lie in [0,1], green proportion}

\item{b}{numeric(1) should lie in [0,1], blue proportion}
}
\description{
function to plot a dot with a given color balance
}
\examples{
par(mfrow=c(2,2))
colordot(1,0,0)
}
